+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [unreleased]
+
+### Deprecated
+
+- Renamed the existing expression parser to `ShuntingYardParser`. This parser
+  is now in maintenance mode and will not receive any new features.
+  `Parser` remains as a deprecated type alias to this legacy implemenation to
+  maintain backwards compatibility.
 
 ## [2.6.0] - 2024-07-31