+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [2.7.0-rc.1] - 2025-02-01
 
 ### Added
 
@@ -235,7 +235,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of standalone version
 
-[unreleased]: https://github.com/fkleon/math-expressions/compare/2.6.0...HEAD
+
+[unreleased]: https://github.com/fkleon/math-expressions/compare/2.7.0...HEAD
+[2.7.0-rc.1]: https://github.com/fkleon/math-expressions/compare/2.6.0...2.7.0-rc.1
 [2.6.0]: https://github.com/fkleon/math-expressions/compare/2.5.0...2.6.0
 [2.5.0]: https://github.com/fkleon/math-expressions/compare/2.4.0...2.5.0
 [2.4.0]: https://github.com/fkleon/math-expressions/compare/2.3.1...2.4.0